Standardized Files
skd 16, condensed newest version of files with final UI rework
