--- State.java
@@ -1,5 +1,0 @@
-public enum State { //enumeration
-    solid,
-    liquid,
-    gas
-}
