+++ nbproject/genfiles.properties
@@ -1,0 +1,8 @@
+build.xml.data.CRC32=97d1d836
+build.xml.script.CRC32=4ef4ac5d
+build.xml.stylesheet.CRC32=f85dc8f2@1.98.0.48
+# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
+# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
+nbproject/build-impl.xml.data.CRC32=97d1d836
+nbproject/build-impl.xml.script.CRC32=ef749105
+nbproject/build-impl.xml.stylesheet.CRC32=d549e5cc@1.98.0.48
