+++ nbproject/private/private.properties
@@ -1,0 +1,2 @@
+compile.on.save=true
+user.properties.file=C:\\Users\\rabra\\AppData\\Roaming\\NetBeans\\12.3\\build.properties
