+++ src/AlchMain/AlchMain.java
@@ -1,0 +1,8 @@
+
+package AlchMain;
+
+public class AlchMain {
+    public static void main(String[] args) {
+    }
+    
+}
