+++ src/AlchMain/GUIMain.java
@@ -1,4 +1,6 @@
-public class AlchMain {
+package GUIMain;
+
+public class GUIMain {
     public static void main(String[] args) {
     }
 }
