+++ src/AlchMain/State.java
@@ -1,0 +1,5 @@
+public enum State { //enumeration
+    solid,
+    liquid,
+    gas
+}
