+++ src/AlchMain/Viewport.java
@@ -1,8 +1,10 @@
+package GUIMain;
+
 import javax.swing.JPanel;
 import java.awt.Graphics;
 
-public class Viewport {
-
+public class Viewport extends JPanel {
+   
     public Viewport(){
         this.setBounds(0,0, 100, 100);
     }
